requirements.txt
app
bcrypt
blinker
click
colorama
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Login
Flask-Mail
gunicorn
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
packaging
protobuf
PyJWT
python-dotenv
Werkzeug



